requirements.txt
streamlit
google-generativeai
python-dotenv
PyMuPDF
python-docx
python-pptx